--- a/src/CriteriaEvaluatorFactory.php
+++ b/src/CriteriaEvaluatorFactory.php
@@ -4,12 +4,18 @@ declare(strict_types=1);
 
 namespace Firehed\Mocktrine;
 
+use function array_key_exists;
+
 /**
  * @template Entity of object
+ *
+ * @internal
  */
 class CriteriaEvaluatorFactory
 {
-    /** @var CriteriaEvaluator[] */
+    /**
+     * @var array<class-string<Entity>, CriteriaEvaluator<Entity>>
+     */
     private static array $instances = [];
 
     /**
@@ -21,6 +27,7 @@ class CriteriaEvaluatorFactory
         if (!array_key_exists($className, self::$instances)) {
             self::$instances[$className] = new CriteriaEvaluator($className);
         }
+        // @phpstan-ignore-next-line (see #3273)
         return self::$instances[$className];
     }
 }