Following the merging of `findBy` and `Selectable` (#17), there was a severe performance regression, where an external test suite that previously took ~32s  degraded to taking ~2:10 (about 4x slower).

These changes not only restore the original performance, but significantly improve it:
- inline entity filtering of `CompositeExpression`s of `TYPE_AND` restored the original performance (which impacts the majority-case of the `findBy()` family)
- re-use of `ReflectionProperty` objects reduced runtime by about 45% on the whole test suite (32s -> 22s), and a ~3x speedup on a test case class that very heavily relies on this library (15s -> 5s)
- switching to a per-class singleton of the criteria evaliator yielded an additional 30% runtime reduction (22s -> 17s) on the whole test suite

Overall, this should be a very, very significant performance improvement (about double from 0.3.0; and nearly an order of magnitude from HEAD) in most test suites, typically with an inconsequential amount of added memory used.